dashbooard keuangan

+++ application/controllers/Admin.php
@@ -50,9 +50,32 @@ class Admin extends CI_Controller
 
     public function hapus_siswa($id)
     {
-        $this->m_model->delete('siswa', 'id_siswa', $id);
-        redirect(base_url('admin/siswa'));
+        $siswa = $this->m_model->get_by_id('siswa', 'id_siswa', $id)->row();
+        if ($siswa) {
+            if ($siswa->foto !== 'User.png') {
+                $file_path = './images/siswa' . $siswa->foto;
+
+                if (file_exists($file_path)) {
+                    if (unlink($file_path)) {
+                        $this->m_model->delete('siswa', 'id_siswa', $id);
+                        redirect(base_url('admin/siswa'));
+                    } else {
+                        echo 'Gagal menghapus file.';
+                    }
+                }
+            } else {
+                $this->m_model->delete('siswa', 'id_siswa', $id);
+                redirect(base_url('admin/siswa'));
+            }
+        } else {
+            echo 'Siswa tidak ditemukan';
+        }
     }
+    // public function hapus_siswa($id)
+    // {
+    //     $this->m_model->delete('siswa', 'id_siswa', $id);
+    //     redirect(base_url('admin/siswa'));
+    // }
 
     public function tambah_siswa()
     {
@@ -93,21 +116,69 @@ class Admin extends CI_Controller
     }
     public function aksi_ubah_siswa()
     {
-        $data = array(
-            'nama_siswa' => $this->input->post('nama'),
-            'nisn' => $this->input->post('nisn'),
-            'gender' => $this->input->post('gender'),
-            'id_kelas' => $this->input->post('id_kelas'),
-        );
+        $foto = $_FILES['foto']['name'];
+        $foto_temp = $_FILES['foto']['tmp_name'];
+
+        // Jika ada foto yang diunggah
+        if ($foto) {
+            $kode = round(microtime(true) * 1000);
+            $file_name = $kode . '_' . $foto;
+            $upload_path = './images/siswa/' . $file_name;
+
+            if (move_uploaded_file($foto_temp, $upload_path)) {
+                // Hapus foto lama jika ada
+                $old_file = $this->m_model->get_siswa_foto_by_id($this->input->post('id_siswa'));
+                if ($old_file && file_exists('./images/siswa/' . $old_file)) {
+                    unlink('./images/siswa/' . $old_file);
+                }
+
+                $data = [
+                    'foto' => $file_name,
+                    'nama_siswa' => $this->input->post('nama'),
+                    'nisn' => $this->input->post('nisn'),
+                    'gender' => $this->input->post('gender'),
+                    'id_kelas' => $this->input->post('kelas'),
+                ];
+            } else {
+                // Gagal mengunggah foto baru
+                redirect(base_url('admin/ubah_siswa/' . $this->input->post('id_siswa')));
+            }
+        } else {
+            // Jika tidak ada foto yang diunggah
+            $data = [
+                'nama_siswa' => $this->input->post('nama'),
+                'nisn' => $this->input->post('nisn'),
+                'gender' => $this->input->post('gender'),
+                'id_kelas' => $this->input->post('kelas'),
+            ];
+        }
+
+        // Eksekusi dengan model ubah_data
         $eksekusi = $this->m_model->ubah_data('siswa', $data, array('id_siswa' => $this->input->post('id_siswa')));
+
         if ($eksekusi) {
-            $this->session->set_flashdata('sukses', 'berhasil');
             redirect(base_url('admin/siswa'));
         } else {
-            $this->session->set_flashdata('error', 'gagal...');
-            redirect(base_url('admin/siswa/ubah_siswa/' . $this->input->post('id_siswa')));
+            redirect(base_url('admin/ubah_siswa/' . $this->input->post('id_siswa')));
         }
     }
+    // public function aksi_ubah_siswa()
+    // {
+    //     $data = array(
+    //         'nama_siswa' => $this->input->post('nama'),
+    //         'nisn' => $this->input->post('nisn'),
+    //         'gender' => $this->input->post('gender'),
+    //         'id_kelas' => $this->input->post('id_kelas'),
+    //     );
+    //     $eksekusi = $this->m_model->ubah_data('siswa', $data, array('id_siswa' => $this->input->post('id_siswa')));
+    //     if ($eksekusi) {
+    //         $this->session->set_flashdata('sukses', 'berhasil');
+    //         redirect(base_url('admin/siswa'));
+    //     } else {
+    //         $this->session->set_flashdata('error', 'gagal...');
+    //         redirect(base_url('admin/siswa/ubah_siswa/' . $this->input->post('id_siswa')));
+    //     }
+    // }
     public function akun()
     {
         $data['user'] = $this->m_model->get_by_id('admin', 'id', $this->session->userdata('id'))->result();
@@ -173,7 +244,8 @@ class Admin extends CI_Controller
         }
     }
 
-    public function dashboard_keuangan() {
+    public function dashboard_keuangan()
+    {
         $this->load->view('keuangan/dashboard');
     }
 }
